+++ lib/haml_lint/configuration_loader.rb
@@ -4,8 +4,8 @@ require 'yaml'
 module HamlLint
   # Manages configuration file loading.
   class ConfigurationLoader
-    DEFAULT_CONFIG_PATH = File.join(HamlLint::HOME, 'config', 'default.yml')
-    CONFIG_FILE_NAME = '.haml-lint.yml'
+    DEFAULT_CONFIG_PATH = File.join(HamlLint::HOME, 'config', 'default.yml').freeze
+    CONFIG_FILE_NAME = '.haml-lint.yml'.freeze
 
     class << self
       # Load configuration file given the current working directory the