+++ lib/haml_lint/linter/unnecessary_string_output.rb
@@ -9,7 +9,7 @@ module HamlLint
   class Linter::UnnecessaryStringOutput < Linter
     include LinterRegistry
 
-    MESSAGE = '`= "..."` should be rewritten as `...`'
+    MESSAGE = '`= "..."` should be rewritten as `...`'.freeze
 
     def visit_tag(node)
       if tag_has_inline_script?(node) && inline_content_is_string?(node)