+++ spec/haml_lint/configuration_loader_spec.rb
@@ -50,7 +50,7 @@ describe HamlLint::ConfigurationLoader do
 
     it 'loads the default config file' do
       described_class.should_receive(:load_from_file)
-        .with(HamlLint::ConfigurationLoader::DEFAULT_CONFIG_PATH)
+                     .with(HamlLint::ConfigurationLoader::DEFAULT_CONFIG_PATH)
       subject
     end
   end
@@ -126,9 +126,8 @@ describe HamlLint::ConfigurationLoader do
       let(:hash) { { 'skip_frontmatter' => true } }
 
       it 'extends the default configuration' do
-        subject.should ==
-          described_class.default_configuration
-                         .merge(HamlLint::Configuration.new(hash))
+        subject.should == described_class.default_configuration
+                          .merge(HamlLint::Configuration.new(hash))
       end
     end
   end