+++ spec/support/directory_spec_helpers.rb
@@ -8,13 +8,14 @@ module DirectorySpecHelpers
   # destroyed at the end of the spec run. Any block passed to this will be
   # executed with the created directory as the working directory.
   #
+  # @yield Executes supplied block in the created directory
   # @return [String] The full path of the directory.
-  def directory(name = 'some-dir', &block)
+  def directory(name = 'some-dir')
     tmpdir = Dir.mktmpdir.tap do |path|
       Dir.chdir(path) do
         Dir.mkdir(name)
         Dir.chdir(name) do
-          block.call if block_given?
+          yield if block_given?
         end
       end
     end